requirements.txt
fastapi[standard]
psycopg2
sqlalchemy
passlib
pyjwt
python-decouple